+++ src/coordinatesForm/form.css
@@ -11,6 +11,9 @@
 }
 
 .upload{
+    width: 100%;
+    height: 90%;
     display: flex;
     flex-direction: column;
+    justify-content: space-evenly;
 }
